refactor: move VRFProof out of crate::json
